# backend/openapi.merged.yaml
openapi: 3.0.3
info:
  title: Kodeme Backend API
  version: 1.0.0
paths:
...
  /organizations:
    post:
      summary: Create a new organization and seed default calendar/settings
      tags:
        - Organizations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - created_by
              properties:
                name:
                  type: string
                subscription_tier:
                  type: string
                created_by:
                  type: string
      responses:
        '200':
          description: Created organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /organizations/{id}:
    get:
      summary: Get organization by id
      tags:
        - Organizations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update organization
      tags:
        - Organizations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                updates:
                  type: object
      responses:
        '200':
          description: Updated organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /organizations/by-user:
    get:
      summary: Get organization for a user
      tags:
        - Organizations
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /settings:
    get:
      summary: Get global settings
      tags:
        - Settings
      responses:
        '200':
          description: Settings object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update global settings
      tags:
        - Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalError'
  /settings/{orgId}/{calendarId}:
    get:
      summary: Get tenant settings for a calendar
      tags:
        - Settings
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
        - name: calendarId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update tenant settings for a calendar
      tags:
        - Settings
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
        - name: calendarId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalError'
components:
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    Organization:
      type: object
      properties:
        org_id:
          type: string
        name:
          type: string
        subscription_tier:
          type: string
        owner_uid:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Settings:
      type: object
      properties:
        org_id:
          type: string
        calendar_id:
          type: string
        workingHours:
          type: object
          properties:
            startTime:
              type: string
            endTime:
              type: string
        workingDays:
          type: array
          items:
            type: integer
        blockedSlots:
          type: array
          items:
            type: object
            additionalProperties: true
        oneOffUnavailableSlots:
          type: array
          items:
            type: object
            additionalProperties: true
        calendarSync:
          type: object
          properties:
            autoCreateEvents:
              type: boolean
            showBusyTimes:
              type: boolean
            syncCancellations:
              type: boolean
        restaurantCity:
          type: string
        restaurantCountry:
          type: string
        restaurantPerimeterKm:
          type: number
        restaurants:
          type: array
          items:
            type: object
            additionalProperties: true
        curatedList:
          type: string
        updatedAt:
          type: string
          format: date-time
...